chore: update the gemspec
The main change here is that we want to change `s.files` so that it does not include all of the generated documentation.

We need to exclude it because otherwise when we build the gem, we get:

```
ERROR:  While executing gem ... (Gem::Package::TooLongFileName)
File "docs/GetVirtualMachineNetworkInterfacesVirtualMachineNetworkInterface200ResponseVirtualMachineNetworkInterface.md"
has a too long name
```

There were also a few hidden files that had been accidentally not committed.