feat: use shorter names for models
This is a new version of the client generated after the apia-open_api gem was updated to produce shorter names for the openapi schema $ref ids.

https://github.com/krystal/apia-openapi/pull/47

Otherwise when we tried to build the Ruby gem we got the error:
`ERROR: While executing gem ... (Gem::Package::TooLongFileName)`

This update also includes other changes introduced by apia-open_api before that version. Specifically more descriptions are included in the schema and thus these are added to the generated documentation.

This is a breaking change but no-one is using this client, so it hasn't been version bumped as a breaking change.